update docs, update release notes, cleanup fields, split fields and move Field to Base, misc bugfixes
TODO:
- fix bug with embedded CompositeField. The fieldname has a duplicate
  when cleaning
