+++ docs/release-notes.md
@@ -10,15 +10,22 @@ hide:
 ### Added
 
 - allow multiple ForeignKeys with the same related_name in one model
+- support ForeignKeys on multiple columns
+- allow selecting columns for ForeignKeys
+- allow skipping creating a ForeignKeyConstraint
 - add ExcludeField for masking fields in submodels
-- add R
+- ConditionalRedirect constant for CompositeField
+- Embeddables via CompositeField
+- add ForeignKeyFactory
 
 ### Changed
 
 - Breaking: Prefetch traversal of foreign keys uses now the foreign key name. For the traversal of RelatedFields everything stays the same.
 - ForeignKeys use now global constraints
-- pk is now a CompositeField (work for multiple primary keys in one model is on the way)
-- ConditionalRedirect constant for CompositeField
+- pk is now a PKField (a variant of the BaseCompositeField)
+- clean and to_columns of BaseField do return empty objects instead of raising NotImplementedError
+- major refactory of ForeignKeys, move logic for single ForeignKeys to subclass
+- move FieldFactory to own file
 
 ## 0.11.1
