+++ edgy/core/db/fields/__init__.py
@@ -11,7 +11,6 @@ from .core import (
     DateTimeField,
     DecimalField,
     EmailField,
-    ExcludeField,
     FloatField,
     IntegerField,
     IPAddressField,
@@ -23,6 +22,7 @@ from .core import (
     URLField,
     UUIDField,
 )
+from .exclude_field import ExcludeField
 from .foreign_keys import ForeignKey
 from .many_to_many import ManyToMany, ManyToManyField
 from .one_to_one_keys import OneToOne, OneToOneField
