+++ edgy/core/db/fields/composite_field.py
@@ -1,17 +1,29 @@
 
 import copy
-from typing import TYPE_CHECKING, Any, Dict, List, Optional, Sequence, Set, Tuple, Type, Union
+from typing import (
+    TYPE_CHECKING,
+    Any,
+    Dict,
+    List,
+    Optional,
+    Sequence,
+    Set,
+    Tuple,
+    Type,
+    Union,
+    cast,
+)
 
 from pydantic import BaseModel
 
 from edgy.core.db.constants import ConditionalRedirect
 from edgy.core.db.fields.base import BaseCompositeField, BaseField
 from edgy.core.db.fields.core import FieldFactory
+from edgy.core.db.fields.exclude_field import ConcreteExcludeField
 from edgy.exceptions import FieldDefinitionError
 
 if TYPE_CHECKING:
-    from edgy.core.db.models.base import EdgyBaseModel
-    from edgy.core.db.models.model import Model
+    from edgy.core.db.models.model import Model, ReflectModel
 
 
 def _removeprefix(text: str, prefix: str) -> str:
@@ -27,12 +39,12 @@ class ConcreteCompositeField(BaseCompositeField):
     Conrete, internal implementation of the CompositeField
     """
 
-    def __init__(self, *, inner_fields:  Union[Sequence[Union[str, Tuple[str, BaseField]]], "EdgyBaseModel", Dict[str, BaseField]] =(), **kwargs: Any):
+    def __init__(self, *, inner_fields:  Union[Sequence[Union[str, Tuple[str, BaseField]]], Type["Model"], Type["ReflectModel"], Dict[str, BaseField]] =(), **kwargs: Any):
         self.inner_field_names: List[str] = []
         self.embedded_field_defs: Dict[str, BaseField] = {}
         if hasattr(inner_fields, "meta"):
-            inner_fields = inner_fields.meta.fields_mapping
             kwargs.setdefault("model", inner_fields)
+            inner_fields = inner_fields.meta.fields_mapping
         if isinstance(inner_fields, dict):
             inner_fields = inner_fields.items()  # type: ignore
         owner = kwargs.pop("owner", None)
@@ -43,13 +55,20 @@ class ConcreteCompositeField(BaseCompositeField):
         for field in inner_fields:
             if isinstance(field, str):
                 self.inner_field_names.append(field)
-            else:
-                field_name = f"{self.prefix_embedded}{field[0]}"
-                field_def = copy.deepcopy(field[1])
-                self.inner_field_names.append(field_name)
-                self.embedded_field_defs[field_name] = field_def
-                # will be overwritten later
-                field_def.owner = owner
+            elif field[1].inherit:
+                # don't copy non inherit fields for excluding PKField and other surprises
+                field_name = field[0]
+                # for preventing suddenly invalid field names
+                if self.prefix_embedded.endswith("_") and field_name.startswith("_"):
+                    raise FieldDefinitionError(f"_ prefixed fields are not supported: {field_name} with prefix ending with _")
+                field_name = f"{self.prefix_embedded}{field_name}"
+                # set field_name and owner
+                # parent is None because not fully initialized
+                field_def = field[1].embed_field(self.prefix_embedded, field_name, owner=owner, parent=None)
+                if field_def is not None:
+                    field_def.exclude = True
+                    self.inner_field_names.append(field_def.name)
+                    self.embedded_field_defs[field_def.name] = field_def
         return super().__init__(
             owner=owner,
             **kwargs,
@@ -60,6 +79,25 @@ class ConcreteCompositeField(BaseCompositeField):
             # PYTHON 3.8 compatibility
             return _removeprefix(name, self.prefix_embedded)
         return name
+
+
+    def embed_field(self, prefix: str, new_fieldname:str, owner: Optional[Union[Type["Model"], Type["ReflectModel"]]]=None, parent: Optional[BaseField]=None) -> BaseField:
+        field_copy = cast(BaseField, super().embed_field(prefix, new_fieldname, owner=owner, parent=self))
+        field_copy.prefix_embedded = f"{prefix}{field_copy.prefix_embedded}"
+        embedded_field_defs = field_copy.embedded_field_defs
+        field_copy.inner_field_names = [f"{prefix}{field_name}" for field_name in field_copy.inner_field_names if field_name not in embedded_field_defs]
+        field_copy.embedded_field_defs = {}
+        for field_name, field in embedded_field_defs.items():
+            # for preventing suddenly invalid field names
+            if self.prefix_embedded.endswith("_") and field_name.startswith("_"):
+                raise FieldDefinitionError(f"_ prefixed fields are not supported: {field_name} with prefix ending with _")
+            field_name = f"{prefix}{field_name}"
+            field_def = field.embed_field(prefix, field_name, owner)
+            if field_def is not None:
+                field_def.exclude = True
+                field_copy.inner_field_names.append(field_def.name)
+                field_copy.embedded_field_defs[field_def.name] = field_def
+        return field_copy
 
     def __get__(self, instance: "Model", owner: Any = None) -> Union[Dict[str, Any], Any]:
         assert len(self.inner_field_names) >= 1
@@ -105,32 +143,35 @@ class ConcreteCompositeField(BaseCompositeField):
         retdict = {}
         if not self.absorb_existing_fields:
             duplicate_fields = set(self.embedded_field_defs.keys()).intersection(
-                {k for k, v in fields_mapping.items() if v.owner is None}
+                {k for k, v in fields_mapping.items() if v.owner is None and not isinstance(v, ConcreteExcludeField)}  # type: ignore
             )
             if duplicate_fields:
                 raise ValueError(f"duplicate fields: {', '.join(duplicate_fields)}")
-            for item in self.embedded_field_defs.items():
+            for field_name, field in self.embedded_field_defs.items():
+                existing_field = fields_mapping.get(field_name, None)
+                if existing_field is not None and existing_field.owner is None and isinstance(existing_field, ConcreteExcludeField):  # type: ignore
+                    continue  # type: ignore
                 # now there should be no collisions anymore
-                cloned_field = copy.copy(item[1])
+                cloned_field = copy.copy(field)
                 # set to the current owner of this field, required in collision checks
                 cloned_field.owner = self.owner
                 cloned_field.inherit = False
-                retdict[item[0]] = cloned_field
+                retdict[field_name] = cloned_field
             return retdict
-        for item in self.embedded_field_defs.items():
-            if item[0] not in fields_mapping:
-                cloned_field = copy.copy(item[1])
+        for field_name, field in self.embedded_field_defs.items():
+            if field_name not in fields_mapping:
+                cloned_field = copy.copy(field)
                 # set to the current owner of this field, required in collision checks
                 cloned_field.owner = self.owner
                 cloned_field.inherit = False
-                retdict[item[0]] = cloned_field
+                retdict[field_name] = cloned_field
             else:
-                absorbed_field = fields_mapping[item[0]]
+                absorbed_field = fields_mapping[field_name]
                 if not getattr(absorbed_field, "skip_absorption_check", False) and not issubclass(
-                    absorbed_field.field_type, item[1].field_type
+                    absorbed_field.field_type, field.field_type
                 ):
                     raise ValueError(
-                        f'absorption failed: field "{item[0]}" handle the type: {absorbed_field.field_type}, required: {item[1].field_type}'
+                        f'absorption failed: field "{field_name}" handle the type: {absorbed_field.field_type}, required: {field.field_type}'
                     )
         return retdict
 
@@ -160,10 +201,10 @@ class CompositeField(FieldFactory):
         inner_fields = kwargs.get("inner_fields")
         if inner_fields is not None:
             if hasattr(inner_fields, "meta"):
-                inner_fields = inner_fields.meta.fields_mapping
                 kwargs.setdefault("model", inner_fields)
+                inner_fields = inner_fields.meta.fields_mapping
             if isinstance(inner_fields, dict):
-                inner_fields = inner_fields.items()  # type: ignore
+                inner_fields = inner_fields.items()
             elif not isinstance(inner_fields, Sequence):
                 raise FieldDefinitionError("inner_fields must be a Sequence, a dict or a model")
             if not inner_fields:
@@ -176,3 +217,6 @@ class CompositeField(FieldFactory):
                 else:
                     if field[0] in inner_field_names:
                         raise FieldDefinitionError(f"duplicate inner field {field}")
+        model = kwargs.get("model", None)
+        if model is not None and not isinstance(model, type):
+            raise FieldDefinitionError(f"model must be type {model}")
