+++ edgy/core/db/relationships/related_field.py
@@ -44,6 +44,10 @@ class RelatedField:
     def foreign_key(self) -> BaseForeignKeyField:
         return cast(BaseForeignKeyField, self.related_from.meta.fields_mapping[self.foreign_key_name])
 
+    @property
+    def is_cross_db(self) -> bool:
+        return self.foreign_key.is_cross_db
+
     def m2m_related(self) -> Any:
         """
         Guarantees the the m2m filter is done by the owner of the call
